move icons into new subfolder and update imports in other components
