+++ packages/react-components/src/components/Footer/Footer.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import SvgBcLogo from "../SvgBcLogo";
+import SvgBcLogo from "../Icons/SvgBcLogo";
 
 import "./Footer.css";
