+++ packages/react-components/src/components/Icons/SvgBcLogo/SvgBcLogo.tsx
