+++ packages/react-components/src/components/Icons/SvgBcLogo/index.ts
