+++ packages/react-components/src/components/Icons/SvgCheckCircleIcon/SvgCheckCircleIcon.tsx
