+++ packages/react-components/src/components/Icons/SvgCheckCircleIcon/index.ts
