+++ packages/react-components/src/components/Icons/SvgCloseIcon/SvgCloseIcon.tsx
