+++ packages/react-components/src/components/Icons/SvgCloseIcon/index.ts
