+++ packages/react-components/src/components/Icons/SvgExclamationCircleIcon/index.ts
