+++ packages/react-components/src/components/Icons/SvgExclamationIcon/SvgExclamationIcon.tsx
