+++ packages/react-components/src/components/Icons/SvgExclamationIcon/index.ts
