+++ packages/react-components/src/components/Icons/SvgInfoIcon/SvgInfoIcon.tsx
