+++ packages/react-components/src/components/Icons/SvgInfoIcon/index.ts
