+++ packages/react-components/src/components/index.ts
@@ -6,12 +6,12 @@ export { default as Header } from "./Header";
 export { default as Footer, FooterLinks } from "./Footer";
 export { default as Form } from "./Form";
 export { default as Select } from "./Select";
-export { default as SvgBcLogo } from "./SvgBcLogo";
-export { default as SvgCheckCircleIcon } from "./SvgCheckCircleIcon";
-export { default as SvgExclamationIcon } from "./SvgExclamationIcon";
-export { default as SvgExclamationCircleIcon } from "./SvgExclamationCircleIcon";
-export { default as SvgCloseIcon } from "./SvgCloseIcon";
-export { default as SvgInfoIcon } from "./SvgInfoIcon";
+export { default as SvgBcLogo } from "./Icons/SvgBcLogo";
+export { default as SvgCheckCircleIcon } from "./Icons/SvgCheckCircleIcon";
+export { default as SvgExclamationIcon } from "./Icons/SvgExclamationIcon";
+export { default as SvgExclamationCircleIcon } from "./Icons/SvgExclamationCircleIcon";
+export { default as SvgCloseIcon } from "./Icons/SvgCloseIcon";
+export { default as SvgInfoIcon } from "./Icons/SvgInfoIcon";
 export { default as Tag } from "./Tag";
 export { default as TagGroup } from "./TagGroup";
 export { default as TagList } from "./TagList";
